Renaming sidebar and delegating more work
