+++ components/EditOverlaySidebar/index.js
@@ -1,0 +1,3 @@
+import EditOverlaySidebar from './EditOverlaySidebar';
+
+export { EditOverlaySidebar };
