--- components/Sidebar/Sidebar.jsx
@@ -1,7 +1,0 @@
-import styles from "./Sidebar.module.scss";
-
-const Sidebar = ({ children }) => {
-  return <div className={styles.Sidebar}>{children}</div>;
-};
-
-export default Sidebar;
