+++ components/index.js
@@ -1,15 +1,8 @@
 export { AddWidgetModal } from './AddWidgetModal';
-
 export { Button } from './Button';
-
 export { Icon } from './Icon';
-
 export { Input } from './Input';
-
 export { Layout } from './Layout';
-
 export { Modal } from './Modal';
-
 export { Overlay } from './Overlay';
-
-export { Sidebar } from './Sidebar';
+export { EditOverlaySidebar } from './EditOverlaySidebar';
